+++ src/components/Header/UserLogo/UserLogo.styled.js
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Icon } from 'components/Icon';
 import { IconPerson } from './ModalUserProfil/ModalUserProfil.styled';
-import { flexBox } from 'styles/mixins';
+
 
 export const UserLogoDiv = styled.div`
     display: flex;
@@ -83,53 +83,5 @@ export const BurgerMenuIcon = styled(Icon)`
     border: none;
     &:hover {
         stroke: ${({ theme }) => theme.colors.primary};
-    }
-`
-export const WrapperToggler = styled.button`
-    display: none;
-    @media ${props => props.theme.media.desktop} {
-        display: block;
-        margin-left: 50px;
-        position: relative;
-        border: transparent;
-
-    }
-    &.active > svg:nth-child(1) {
-        fill: ${({ theme }) => theme.colors.primary};
-        }
-    &.active > svg:nth-child(2) {
-            transform: translateX(35px);
-        }
-`
-export const Toggler = styled(Icon)`
-    display: none;
-    @media ${props => props.theme.media.desktop} {
-        display: block;
-        width: 62px;
-        height: 27px;
-        fill: #ffffff;
-        background: #EFEFEF;
-        box-shadow: inset 0px 6px 8px 3px rgba(0, 0, 0, 0.1);
-        border-radius: 50px;
-        border: transparent;
-
-        &:hover {
-        fill: #EBF3D4;
-    }
-    }
-`
-export const TogglerBtn = styled(Icon)`
-    display: none;
-    @media ${props => props.theme.media.desktop} {
-        display: block;
-        position: absolute;
-        top: 3px;
-        left: 3px;
-        transition: transform 0.5s ease-in;
-        width: 25px;
-        height: 25px;
-        background: linear-gradient(180deg, #FFFFFF 0%, #E8EAEA 100%);
-        border-radius: 50%;
-        fill: #FFFFFF;
     }
 `
